Enum verification::Error

pub enum Error {
    Duplicate,
    DuplicatedTransactions,
    Empty,
    Pow,
    FuturisticTimestamp,
    Timestamp,
    Coinbase,
    Transaction(usize, TransactionError),
    Difficulty {
        expected: Compact,
        actual: Compact,
    },
    MerkleRoot,
    CoinbaseOverspend {
        expected_max: u64,
        actual: u64,
    },
    CoinbaseScript,
    MaximumSigops,
    MaximumSigopsCost,
    CoinbaseSignatureLength(usize),
    Size(usize),
    Weight,
    NonFinalBlock,
    OldVersionBlock,
    TransactionFeeAndRewardOverflow,
    TransactionFeesOverflow,
    ReferencedInputsSumOverflow,
    WitnessInvalidNonceSize,
    WitnessMerkleCommitmentMismatch,
    UnexpectedWitness,
    NonCanonicalTransactionOrdering,
    Database(DBError),
}

All possible verification errors

Variants
Duplicate

has an equal duplicate in the chain

DuplicatedTransactions

Contains duplicated transactions

Empty

No transactions in block

Pow

Invalid proof-of-work (Block hash does not satisfy nBits)

FuturisticTimestamp

Futuristic timestamp

Timestamp

Invalid timestamp

Coinbase

First transaction is not a coinbase transaction

Transaction(usize, TransactionError)

One of the transactions is invalid (corresponding index and specific transaction error)

Difficulty

nBits do not match difficulty rules

Fields of Difficulty

expected: Compact

actual: Compact

MerkleRoot

Invalid merkle root

CoinbaseOverspend

Coinbase spends too much

Fields of CoinbaseOverspend

expected_max: u64

actual: u64

CoinbaseScript

Coinbase has invalid script sig prefix (BIP90 -> BIP34)

MaximumSigops

Maximum sigops operations exceeded - will not provide how much it was in total since it stops counting once MAX_BLOCK_SIGOPS is reached

MaximumSigopsCost

Maximum sigops operations cost exceeded

CoinbaseSignatureLength(usize)

Coinbase signature is not in the range 2-100

Size(usize)

Block size is invalid

Weight

Block weight is invalid

NonFinalBlock

Block transactions are not final.

OldVersionBlock

Old version block.

TransactionFeeAndRewardOverflow

Sum of the transaction fees in block + coinbase reward exceeds u64::max

TransactionFeesOverflow

Sum of the transaction fees in block exceeds u64::max

ReferencedInputsSumOverflow

Sum of all referenced outputs in block transactions resulted in the overflow

WitnessInvalidNonceSize

SegWit: bad witess nonce size

WitnessMerkleCommitmentMismatch

SegWit: witness merkle mismatch

UnexpectedWitness

SegWit: unexpected witness

NonCanonicalTransactionOrdering

Non-canonical tranasctions ordering within block

Database(DBError)

Database error