Enum verification::TransactionError

pub enum TransactionError {
    Empty,
    NullNonCoinbase,
    CoinbaseSignatureLength(usize),
    MaxSize,
    MaxSigops,
    MemoryPoolCoinbase,
    Input(usize),
    Maturity,
    Signature(usize),
    UnknownReference(H256),
    Overspend,
    SignatureMallformed(String),
    Sigops(usize),
    SigopsP2SH(usize),
    MisplacedCoinbase,
    UnspentTransactionWithTheSameHash,
    UsingSpentOutput(H256, u32),
}

Possible transactions verification errors

Variants
Empty

Transaction has no inputs or no outputs

NullNonCoinbase

Transaction is not coinbase transaction but has null inputs

CoinbaseSignatureLength(usize)

Coinbase signature is not in the range 2-100

MaxSize

Transaction size exceeds block size limit

MaxSigops

Transaction has more sigops than it's allowed

MemoryPoolCoinbase

Transaction is a part of memory pool, but is a coinbase

Input(usize)

Not found corresponding output for transaction input

Maturity

Referenced coinbase output for the transaction input is not mature enough

Signature(usize)

Signature invalid for given input

UnknownReference(H256)

Unknown previous transaction referenced

Overspend

Spends more than claims

SignatureMallformed(String)

Signature script can't be properly parsed

Sigops(usize)

Too many signature operations

SigopsP2SH(usize)

Too many signature operations once p2sh operations included

MisplacedCoinbase

Coinbase transaction is found at position that is not 0

UnspentTransactionWithTheSameHash

Not fully spent transaction with the same hash already exists, bip30.

UsingSpentOutput(H256, u32)

Using output that is surely spent